Original 4.3BSD-Quasijarus0a release notes
4.3BSD-Quasijarus0a was released on 1999-10-10.

As you can read on the Quasijarus Project History page, this release does what 4.3BSD-Quasijarus0 was intended but unfortunately failed to do: provide a demark between CSRG and Quasijarus. It is the UNIX operating system in its state at the point of its transition from CSRG to Quasijarus. In preparing this release, I have sweated for 9 months to reconstruct, stabilize, shape up, and package up the system I have inherited from CSRG, but very carefully avoided any changes of my own. In particular, the VAX hardware support in this release consists of all CPUs and devices that CSRG supported or tried to support. This support has been highly polished, i.e., all of it works as expected and as advertised, which at CSRG was not the case in some areas. However, no all-new hardware support has been added. This release has been carefully polished and shined for both users and developers. For users I have been extremely careful to make this release of full production quality not less than that of the very original 4.3BSD. In the system as distributed, the system header files are copied to /usr/include rather than symlinked, so that a production site can run without any sources installed and still give users the full functionality, including the ability to compile their own programs. For developers, the source tree has been fully SCCS-reconstructed, polished, and shined. The system fully rebuilds from the source like a charm.

4.3BSD-Quasijarus0 and 4.3BSD-Quasijarus0a contain a serious bug in c2, the peephole optimiser, which causes it under certain circumstances to throw out instructions loading registers which are later used in deferred or indexed addressing modes. The resulting misoptimised code thus dereferences bogus addresses, dumping core or even worse, behaving unpredictably. Since all system binaries are compiled with -O this potentially affects every binary in the system. This bug has been fixed in 4.3BSD-Quasijarus0b.

@(#)0a.html 1.1 03/12/18

Michael Sokolov
msokolov@ivan.Harhan.ORG